+++ py/desitarget/test/test_priorities.py
@@ -159,7 +159,7 @@ class TestPriorities(unittest.TestCase):
     def test_mask_priorities(self):
         for mask in [desi_mask, bgs_mask, mws_mask]:
             for name in mask.names():
-                if name.startswith('STD') or name in ['BGS_ANY', 'MWS_ANY', 'SECONDARY_ANY',
+                if name.startswith('STD') or name in ['BGS_ANY', 'MWS_ANY', 'SCND_ANY',
                                                       'IN_BRIGHT_OBJECT', 'NEAR_BRIGHT_OBJECT',
                                                       'BRIGHT_OBJECT', 'SKY', 'SV', 'NO_TARGET']:
                     self.assertEqual(mask[name].priorities, {}, 'mask.{} has priorities?'.format(name))