+++ doc/changes.rst
@@ -5,7 +5,25 @@ desitarget Change Log
 0.30.2 (unreleased)
 -------------------
 
-* No changes yet.
+* First implementation for secondary targets [`PR #507`_]. Includes:
+    * Framework and design for secondary targeting process.
+    * Works automatically for both Main Survey and SV files.
+    * New bitmasks for secondaries that populate ``SCND_TARGET`` column.
+        * can have any ``PRIORITY_INIT`` and ``NUMOBS_INIT``.
+    * A reserved "veto" bit to categorically reject targets.
+    * Rigorous checking of file formats...
+        * ...and that files correspond to secondary bits.
+    * Example files and file structure (at NERSC) in ``SCND_DIR``.
+        * /project/projectdirs/desi/target/secondary.
+    * Secondary targets are matched to primary targets on RA/Dec.
+        * unless a (per-source) ``OVERRIDE`` column is set to ``True``.
+    * Secondary-primary matches share the primary ``TARGETID``.
+    * Non-matches and overrides have their own ``TARGETID``.
+        * with ``RELEASE == 0``.
+    * Non-override secondary targets are also matched to themselves.
+        * ``TARGETID`` and ``SCND_TARGET`` correspond for matches.
+
+.. _`PR #507`: https://github.com/desihub/desitarget/pull/507
 
 0.30.1 (2019-06-18)
 -------------------
@@ -15,7 +33,7 @@ desitarget Change Log
 * Follow-up PR to `PR #496`_ with two changes and bug fixes [`PR #505`_]:
     * Select QSO targets using random forest by default.
     * Bug fix: Correctly populate ``REF_CAT`` column (needed to correctly set
-      MWS targeting bits).     
+      MWS targeting bits).
 
 .. _`PR #505`: https://github.com/desihub/desitarget/pull/505
 .. _`PR #506`: https://github.com/desihub/desitarget/pull/506
@@ -94,6 +112,7 @@ desitarget Change Log
 
 0.29.0 (2019-03-22)
 -------------------
+
 * Update SV selection for DR8 [`PR #477`_]. Includes:
     * New SV targeting bits for QSOs and LRGs.
     * New SV selection algorithms for QSOs, ELGs and LRGs.
@@ -395,7 +414,6 @@ bit names and selection function names.
 
 .. _`PR #334`: https://github.com/desihub/desitarget/pull/334
 
-
 0.21.0 (2018-07-18)
 -------------------
 
@@ -464,7 +482,6 @@ bit names and selection function names.
 .. _`PR #329`: https://github.com/desihub/desitarget/pull/329
 .. _`PR #331`: https://github.com/desihub/desitarget/pull/331
 .. _`PR #332`: https://github.com/desihub/desitarget/pull/332
-
 
 0.20.1 (2018-03-29)
 -------------------